refactor: unite typedef and types
