+++ interface/static.ts
@@ -1,5 +1,5 @@
 import { v4 as uuidv4 } from "uuid";
-import { Country } from "../types/common";
+import { Country } from "../typedef/common";
 import {
   CategoryTitle,
   ReviewCategoryFilterListType,
