+++ src/api/user/api.ts
@@ -1,5 +1,5 @@
-import { LanguageType, ReviewSortType } from "@/../types/common";
-import { MemberReviewsRequest } from "@/../types/member/request";
+import { LanguageType, ReviewSortType } from "../../../typedef/common";
+import { MemberReviewsRequest } from "../../../typedef/member/request";
 import axios from "axios";
 import { BeerSortEnum, OAUTH_PROVIDER } from "../../../interface/types";
 import {
